♻️ Refactor window utils
