+++ src/app/background.tsx
@@ -1,14 +1,11 @@
 import { startCaptureHighlights } from './utils/highlights';
 import { waitForOverwolf } from './utils/overwolf';
+import { restoreWindow, WINDOWS } from './utils/windows';
 
 console.log('Starting background process');
-waitForOverwolf().then(() => {
-  // overwolf.windows.obtainDeclaredWindow('desktop', (result) => {
-  //   overwolf.windows.restore(result.window.id, console.log);
-  // });
+waitForOverwolf().then(async () => {
   startCaptureHighlights();
 
-  overwolf.windows.obtainDeclaredWindow('development', (result) => {
-    overwolf.windows.restore(result.window.id, console.log);
-  });
+  restoreWindow(WINDOWS.DEVELOPMENT);
+  restoreWindow(WINDOWS.DESKTOP);
 });
